+++ web/client-js/src/store/board/types.ts
@@ -25,7 +25,16 @@
  */
 
 /**
- * An alias for the {@code proto.spine_examples.kanban.BoardView} event.
+ * Module name.
+ */
+export const BOARD = "BOARD";
+
+/**
+ * Aliases for frequently used types.
+ */
+
+/**
+ * An alias for the {@code proto.spine_examples.kanban.BoardView} projection.
  */
 export type Board = proto.spine_examples.kanban.BoardView;
 
@@ -40,14 +49,14 @@ export type BoardCreated = proto.spine_examples.kanban.BoardCreated;
 export type ColumnAdded = proto.spine_examples.kanban.ColumnAdded;
 
 /**
- * The local state of the Kanban web application.
+ * The local state of the Kanban board.
  */
-export interface KanbanState {
+export interface BoardState {
   board: Board | null;
 }
 
 /**
- * Exposes mutations of the local {@plain KanbanState state}.
+ * Exposes mutations of the local {@linkplain BoardState board state}.
  */
 export const Mutation = {
   /**
@@ -63,8 +72,7 @@ export const Mutation = {
 };
 
 /**
- * Exposes possible interactions with the Kanban web server, e.g.
- * send a command and subscribe to produced events.
+ * Exposes interactions with the remote board state.
  */
 export const Action = {
   /**
