+++ web/client-js/src/store/root/types.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2000-2021 TeamDev. All rights reserved.
+ * TeamDev PROPRIETARY and CONFIDENTIAL.
+ * Use is subject to license terms.
+ */
+
+/**
+ * The root state of the Vuex store.
+ */
+// It is required to have a typed root state.
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface RootState {}
